requirements.txt
allure-pytest~=2.14.3
selenium~=4.33.0
allure-python-commons~=2.14.3
pytest~=8.3.5